Split tests across two files
