+++ project/api/tests/__init__.py
@@ -1,0 +1,2 @@
+from .model_tests import *
+from .api_tests import *
